+++ amazonorders/entity/order.py
@@ -92,8 +92,12 @@ class Order(Parsable):
         #: The Order coupon savings. Only populated when ``full_details`` is ``True``.
         self.coupon_savings: Optional[float] = self._if_full_details(
             self._parse_currency("coupon", combine_multiple=True))
+        #: The Order reward points. Only populated when ``full_details`` is ``True``.
+        self.reward_points: Optional[float] = self._if_full_details(
+            self._parse_currency("reward", combine_multiple=True))
         #: The Order Subscribe & Save discount. Only populated when ``full_details`` is ``True``.
-        self.subscription_discount: Optional[float] = self._if_full_details(self._parse_currency("subscribe"))
+        self.subscription_discount: Optional[float] = (self._if_full_details(self._parse_currency("subscribe"))
+                                                       or self._if_full_details(self._parse_currency("subscription")))
         #: The Order total before tax. Only populated when ``full_details`` is ``True``.
         self.total_before_tax: Optional[float] = self._if_full_details(self._parse_currency("before tax"))
         #: The Order estimated tax. Only populated when ``full_details`` is ``True``.